Update README.md
This package of systems analysis tools includes desktop applications for implementing simulation, optimization and multi-objective analysis. The package is developed as part of the "Managing Water Resources-Methods and Applications for a Systems Approach" textbook that can be downloaded from www.slobodansimonovic.com. The state-of-the-art simulation software Vensim PLE (Personal Learning Edition) is enclosed for the implementation of system dynamics simulation. This program was developed by Ventana Systems, which has kindly given permission for its use in this context. It can be downloaded directly from the Ventana web site https://www.ventanasystems.com/software/. The Systems Analysis Tools package includes seven more original computer programs developed in the user-friendly Windows environment. They are:

LINPRO - a linear programming optimization tool;
FUZZYLINPRO - a program for the implementation of fuzzy linear programming optimization;
EVOLPRO - for the implementation of evolutionary optimization;
COMPRO - for the implementation of the deterministic multi-objective analysis tool of compromise programming;
FUZZYCOMPRO - which implements fuzzy compromise programming, for multi-objective analysis under uncertainty;
FUZZYCOMPROGDM - for the support of group decision making using fuzzy compromise programming under uncertainty;
SUSTAINPRO - a package of four programs for the computation of fairness, risk, reversibility and consensus sustainability criteria.
You can download all the programs from this website. 

Each program is presented in the same way, with:
- read.me file with installation instructions;
- folder containing the main program files;
- folder containing all the examples presented in the textbook.
- Vensim PLE is accompanied with a short tutorial developed by Craig W. Kirkwood, of Arizona State University. I am grateful to the author for permission to provide it here. The other seven programs have very extensive help manuals, which are integrated into the Windows environment. These provide detailed instructions on program use, data preparation, data import and interpretation of the results.